removed debug statement
